Try multiple bounding boxes in TestGenerator.
Try bounding boxes of decreasing width/increasing height, starting
with a bounding box that is guaranteed to fit all rectangle side-by-side.
Iteratively decrease the width or increase the height when the rectangles can
resp. can't be packed into the current bounding box, and maintain a reference
to the best bounding box.